+++ tomcat/src/main/java/org/apache/coyote/controller/PageController.java
@@ -1,0 +1,38 @@
+package org.apache.coyote.controller;
+
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.nio.file.Files;
+
+import static org.apache.coyote.http11.Status.OK;
+
+public class PageController extends AbstractController {
+
+    @Override
+    protected HttpResponse doGet(HttpRequest httpRequest) throws IOException {
+        HttpResponse httpResponse = new HttpResponse();
+
+        String responseBody = getResource(httpRequest.getPath());
+
+        httpResponse.setStatusLine(OK);
+        httpResponse.setContentType(httpRequest.getContentType());
+        httpResponse.setResponseBody(responseBody);
+        httpResponse.setContentLength(responseBody.getBytes().length);
+
+        return httpResponse;
+    }
+
+    @Override
+    protected HttpResponse doPost(HttpRequest httpRequest) {
+        return null;
+    }
+
+    private String getResource(String path) throws IOException {
+        URL resource = getClass().getClassLoader().getResource("static" + path);
+        File file = new File(resource.getFile());
+        return new String(Files.readAllBytes(file.toPath()));
+    }
+}
